+++ chanjo/init/__init__.py
@@ -1,0 +1,3 @@
+"""
+Bootstrap a new instance of chanjo.
+"""