+++ chanjo/parse/sambamba/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from .parse import depth_output, expand_header, expand_row