+++ tests/store/test_store_models.py
@@ -1,14 +1,12 @@
 # -*- coding: utf-8 -*-
-from chanjo.store import Exon, Exon_Transcript
+from chanjo.store.models import TranscriptStat, Exon
 
 
-def test_Exon_Transcript():
-    # test table name
-    assert Exon_Transcript.name == 'exon__transcript'
-
-
-def test_Interval():
-    # set up
-    exon = Exon(exon_id='exon1', chromosome='chr1', start=10, end=1000)
-
-    assert len(exon) == 991
+def test_TranscriptStat():
+    # GIVEN a transcript stat with two incomplete exons
+    exons = [Exon('1', 10, 100, 99.1), Exon('1', 200, 300, 80.5)]
+    stat = TranscriptStat(mean_coverage=10.3, incomplete_exons=exons)
+    # WHEN accessing them
+    parsed_exons = list(stat.incomplete_exons)
+    # THEN should be the same
+    assert parsed_exons == exons